rename more path references, part 2
