+++ .prettierignore
@@ -1,5 +1,6 @@
 node_modules/
 generated/
+generated-icons/
 lib/
 dist/
 /site/
